Native Redis Session Storage update
