+++ src/Symfony/Component/HttpFoundation/Session/Storage/Handler/NativeRedisSessionHandler.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\HttpFoundation\Session\Storage;
+namespace Symfony\Component\HttpFoundation\Session\Storage\Handler;
 
 /**
  * NativeRedisSessionStorage.
@@ -20,37 +20,24 @@ namespace Symfony\Component\HttpFoundation\Session\Storage;
  *
  * @author Andrej Hudec <pulzarraider@gmail.com>
  */
-class NativeRedisSessionStorage extends AbstractSessionStorage
+class NativeRedisSessionHandler extends NativeSessionHandler
 {
-    /**
-     * @var string
-     */
-    private $savePath;
-
     /**
      * Constructor.
      *
      * @param string $savePath Path of redis server.
-     * @param array  $options  Session configuration options.
-     *
-     * @see AbstractSessionStorage::__construct()
      */
-    public function __construct($savePath = 'tcp://127.0.0.1:6379?persistent=0', array $options = array())
+    public function __construct($savePath = 'tcp://127.0.0.1:6379?persistent=0')
     {
         if (!extension_loaded('redis')) {
             throw new \RuntimeException('PHP does not have "redis" session module registered');
         }
 
-        $this->savePath = $savePath;
-        parent::__construct($options);
-    }
+        if (null === $savePath) {
+            $savePath = ini_get('session.save_path');
+        }
 
-    /**
-     * {@inheritdoc}
-     */
-    protected function registerSaveHandlers()
-    {
         ini_set('session.save_handler', 'redis');
-        ini_set('session.save_path', $this->savePath);
+        ini_set('session.save_path', $savePath);
     }
 }
